+++ pages/404.js
@@ -8,11 +8,11 @@ export default function Custom404() {
 					4<span className="text-indigo-200">o</span>4
 				</h1>
 				<div>
-					<h2 className="text-3xl pt-12">Oooppsie!</h2>
+					<h2 className="text-3xl pt-12">Oopsie!</h2>
 					<p className="pt-8 max-w-[500px] mx-auto">
 						The page you're trying to reach has been moved or doesn't exist.
-						Restarting from <Link href="/">Home</Link> page and you will
-						undoubtedly fix the issue.
+						Restarting from <Link href="/">Home</Link> and you will undoubtedly
+						fix the issue.
 					</p>
 				</div>
 			</div>