+++ pages/dashboard/user/index.js
@@ -7,7 +7,7 @@ function User(props) {
 	const { data: session } = useSession();
 	const router = useRouter();
 
-	const submitForm = e => {
+	const submitForm = (e) => {
 		e.preventDefault();
 
 		const firstName = e.target.firstName.value;
@@ -53,24 +53,35 @@ function User(props) {
 		};
 
 		fetch(props.SITE_URI + "/api/users", requestOptions)
-			.then(response => response.json())
-			.then(result => {
+			.then((response) => response.json())
+			.then((result) => {
 				if (result.success === true) {
 					router.reload(window.location.pathname);
 				}
 			})
-			.catch(error => console.log("error", error));
+			.catch((error) => console.log("error", error));
 	};
 
 	return (
-		<div className="mt-10 max-w-7xl mx-auto px-4 md:px-6">
+		<div className="mt-10 max-w-7xl mx-auto px-4 md:px-6 text-white flex flex-col">
 			<Head>
 				<title>Profile</title>
 				<meta name="viewport" content="initial-scale=1.0, width=device-width" />
 			</Head>
-			<div className="block mb-2 dark:text-white">User</div>
+
+			<div className="mb-10 flex items-center justify-between border-b border-white pb-8">
+				{" "}
+				<h3 className="block text-white text-2xl">User Information</h3>
+				<button
+					className="py-2 bg-white text-black self-end rounded-lg w-36 hover:bg-black hover:text-white"
+					onClick={() => signOut()}
+				>
+					Log out
+				</button>
+			</div>
+
 			<form
-				className="w-full max-w-lg"
+				className="lg:ml-0 w-full px-8 lg:px-12 md:px-0 md:w-[60vw] flex flex-col mx-auto"
 				method="POST"
 				action=""
 				onSubmit={submitForm}
@@ -258,18 +269,19 @@ function User(props) {
 				</div>
 
 				<button
-					className="text-white bg-blue-700 hover:bg-blue-800 focus:ring-4 focus:outline-none focus:ring-blue-300  rounded-lg  w-full sm:w-auto px-5 py-2.5 text-center dark:bg-blue-600 dark:hover:bg-blue-700 dark:focus:ring-blue-800"
+					className="mt-6 w-36 text-white bg-blue-700 hover:bg-blue-800 focus:ring-4 focus:outline-none focus:ring-blue-300 rounded-lg px-5 py-2.5 text-center dark:bg-blue-600 dark:hover:bg-blue-700 dark:focus:ring-blue-800"
 					type="submit"
 				>
 					Save
 				</button>
+
+				<div
+					className="hover:cursor-pointer my-10 text-white border-t border-b border-white py-4 "
+					href="#"
+				>
+					<p className="text-red-600">Delete Account</p>
+				</div>
 			</form>
-			<Link
-				className="text-white bg-red-700 hover:bg-red-800 focus:ring-4 focus:outline-none focus:ring-red-300 rounded-lg w-full sm:w-auto px-5 py-2.5 text-center dark:bg-red-600 dark:hover:bg-red-700 dark:focus:ring-red-800"
-				href="#"
-			>
-				Delete Account
-			</Link>
 		</div>
 	);
 }
@@ -291,11 +303,11 @@ export async function getServerSideProps(context) {
 			process.env.SITE_URI + "/api/users?email=" + session._doc.email,
 			requestOptions
 		)
-			.then(response => response.json())
-			.then(result => {
+			.then((response) => response.json())
+			.then((result) => {
 				return result.data[0];
 			})
-			.catch(error => console.log("error", error));
+			.catch((error) => console.log("error", error));
 
 		response.SITE_URI = process.env.SITE_URI;
 		return { props: response };